初期完成

+++ src/main/java/com/example/bookstore_project/service/impl/BookStoreServiceImpl.java
@@ -3,16 +3,15 @@ package com.example.bookstore_project.service.impl;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import com.example.bookstore_project.service.ifs.BookStoreService;
 import com.example.bookstore_project.vo.BookRankRes;
-import com.example.bookstore_project.vo.BookStoreReq;
 import com.example.bookstore_project.vo.BookStoreRes;
 import com.example.bookstore_project.vo.orderBookReq;
 import com.example.bookstore_project.constans.BookStoreRtnCode;
@@ -48,6 +47,9 @@ public class BookStoreServiceImpl implements BookStoreService{
 
 	@Override
 	public BookStoreRes deleteById(String id) {
+		if(!bookStoredao.findById(id).isPresent()) {
+			return new BookStoreRes(BookStoreRtnCode.ID_REQUIRED.getMessage());
+		}
 		
 		 bookStoredao.deleteById(id);
 		 
@@ -103,6 +105,7 @@ public class BookStoreServiceImpl implements BookStoreService{
 		List<String>buy = new ArrayList<>();
 		buy.add("潦禦灿:");
 		messageList.setMessagelist(buy);
+		
 		for (orderBookReq entry : orderList) {
 			if(entry.getId().isEmpty()){
 				return new BookStoreRes(null,BookStoreRtnCode.ID_REQUIRED.getMessage());
@@ -130,4 +133,160 @@ public class BookStoreServiceImpl implements BookStoreService{
 		messageList.messagelist.add("羆肂:"+totalprice);
 		return messageList;
 	}
+
+	@Override
+	public BookStoreRes searchBooks(String code,String id_Or_Name_Or_Writer) {
+		List<String>message = new ArrayList<>();
+		message.add("琩高挡狦: ");
+		BookStoreRes messageList = new BookStoreRes();
+		messageList.setMessagelist(message);
+		
+		if(!code.equalsIgnoreCase("0000")) {
+		if(id_Or_Name_Or_Writer.isEmpty()) {
+			return new BookStoreRes(null,BookStoreRtnCode.ID_NAME_WRITER_REQUIRED.getMessage());
+		}
+		else if(bookStoredao.findById(id_Or_Name_Or_Writer).isPresent()) {
+			Optional<BookStore> book = bookStoredao.findById(id_Or_Name_Or_Writer);
+			messageList.messagelist.add("ID:"+book.get().getId());
+			messageList.messagelist.add(":"+book.get().getName());
+			messageList.messagelist.add(":"+book.get().getWriter());
+			messageList.messagelist.add("基:"+book.get().getPrice());
+		}
+        else if(!bookStoredao.findByName(id_Or_Name_Or_Writer).isEmpty()) {
+        	 List<BookStore> books = bookStoredao.findByName(id_Or_Name_Or_Writer);
+        	 for(BookStore info : books) {
+        		messageList.messagelist.add("ID:"+info.getId());
+     			messageList.messagelist.add(":"+info.getName());
+     			messageList.messagelist.add(":"+info.getWriter());
+     			messageList.messagelist.add("基:"+info.getPrice());
+     			messageList.messagelist.add("=============");
+        	 }
+		}
+        else if(!bookStoredao.findByWriter(id_Or_Name_Or_Writer).isEmpty()) {
+        	List<BookStore> books = bookStoredao.findByWriter(id_Or_Name_Or_Writer);
+       	    for(BookStore info : books) {
+       		    messageList.messagelist.add("ID:"+info.getId());
+    			messageList.messagelist.add(":"+info.getName());
+    			messageList.messagelist.add(":"+info.getWriter());
+    			messageList.messagelist.add("基:"+info.getPrice());
+    			messageList.messagelist.add("=============");
+       	 }
+		}
+        else {
+        	return new BookStoreRes(null,BookStoreRtnCode.ID_NAME_WRITER_REQUIRED.getMessage());
+        }
+	}
+		if(code.equalsIgnoreCase("0000")) {
+			if(id_Or_Name_Or_Writer.isEmpty()) {
+				return new BookStoreRes(null,BookStoreRtnCode.ID_NAME_WRITER_REQUIRED.getMessage());
+			}
+			else if(bookStoredao.findById(id_Or_Name_Or_Writer).isPresent()) {
+				Optional<BookStore> book = bookStoredao.findById(id_Or_Name_Or_Writer);
+				messageList.messagelist.add("ID:"+book.get().getId());
+				messageList.messagelist.add(":"+book.get().getName());
+				messageList.messagelist.add(":"+book.get().getWriter());
+				messageList.messagelist.add("基:"+book.get().getPrice());
+				messageList.messagelist.add("畐:"+book.get().getStorage());
+				messageList.messagelist.add("綪扳秖:"+book.get().getSales());
+			}
+	        else if(!bookStoredao.findByName(id_Or_Name_Or_Writer).isEmpty()) {
+	        	 List<BookStore> books = bookStoredao.findByName(id_Or_Name_Or_Writer);
+	        	 for(BookStore info : books) {
+	        		messageList.messagelist.add("ID:"+info.getId());
+	     			messageList.messagelist.add(":"+info.getName());
+	     			messageList.messagelist.add(":"+info.getWriter());
+	     			messageList.messagelist.add("基:"+info.getPrice());
+	     			messageList.messagelist.add("畐:"+info.getStorage());
+					messageList.messagelist.add("綪扳秖:"+info.getSales());
+	     			messageList.messagelist.add("=============");
+	        	 }
+			}
+	        else if(!bookStoredao.findByWriter(id_Or_Name_Or_Writer).isEmpty()) {
+	        	List<BookStore> books = bookStoredao.findByWriter(id_Or_Name_Or_Writer);
+	       	    for(BookStore info : books) {
+	       		    messageList.messagelist.add("ID:"+info.getId());
+	    			messageList.messagelist.add(":"+info.getName());
+	    			messageList.messagelist.add(":"+info.getWriter());
+	    			messageList.messagelist.add("基:"+info.getPrice());
+	    			messageList.messagelist.add("畐:"+info.getStorage());
+					messageList.messagelist.add("綪扳秖:"+info.getSales());
+	    			messageList.messagelist.add("=============");
+	       	 }
+			}
+	        else {
+	        	return new BookStoreRes(null,BookStoreRtnCode.ID_NAME_WRITER_REQUIRED.getMessage());
+	        }
+		}
+		return messageList;
+	}
+
+	@Override
+	public BookStoreRes updateStorage(String code, String id, int num) {
+		if(!code.equalsIgnoreCase("0000")) {
+			return new BookStoreRes(null,BookStoreRtnCode.IDPOWER_EXISTED.getMessage());
+		}
+		if(!StringUtils.hasText(id)) {
+			return new BookStoreRes(null,BookStoreRtnCode.ID_REQUIRED.getMessage());
+		}
+		List<String>message = new ArrayList<>();
+		message.add("穝挡狦:");
+		BookStoreRes messageList = new BookStoreRes();
+		messageList.setMessagelist(message);
+		
+		if(bookStoredao.findById(id).isPresent()) {
+			Optional<BookStore> book = bookStoredao.findById(id);
+			if(book.get().getStorage() + num < 0) {
+				messageList.messagelist.add(BookStoreRtnCode.STORAGE_REQUIRED.getMessage());;
+				messageList.messagelist.add("ヘ玡畐:"+book.get().getStorage());
+				return messageList;
+			}
+			messageList.messagelist.add("穝玡畐:"+book.get().getStorage());
+			messageList.messagelist.add("===========");
+			messageList.messagelist.add("穝:");
+			book.get().setStorage(book.get().getStorage() + num);
+			bookStoredao.save(book.get());
+			messageList.messagelist.add("ID:"+book.get().getId());
+			messageList.messagelist.add(":"+book.get().getName());
+			messageList.messagelist.add(":"+book.get().getWriter());
+			messageList.messagelist.add("基:"+book.get().getPrice());
+			messageList.messagelist.add("畐:"+book.get().getStorage());
+		}
+		return messageList;
+	}
+
+	@Override
+	public BookStoreRes updatePrice(String code, String id, int price) {
+		if(!code.equalsIgnoreCase("0000")) {
+			return new BookStoreRes(null,BookStoreRtnCode.IDPOWER_EXISTED.getMessage());
+		}
+		if(!StringUtils.hasText(id)) {
+			return new BookStoreRes(null,BookStoreRtnCode.ID_REQUIRED.getMessage());
+		}
+		List<String>message = new ArrayList<>();
+		message.add("穝挡狦:");
+		BookStoreRes messageList = new BookStoreRes();
+		messageList.setMessagelist(message);
+		
+		if(bookStoredao.findById(id).isPresent()) {
+			Optional<BookStore> book = bookStoredao.findById(id);
+			if(price < 0) {
+				messageList.messagelist.add(BookStoreRtnCode.PRICE_REQUIRED.getMessage());;
+				messageList.messagelist.add("ヘ玡基:"+book.get().getPrice());
+				return messageList;
+			}
+			messageList.messagelist.add("穝玡基:"+book.get().getPrice());
+			messageList.messagelist.add("===========");
+			messageList.messagelist.add("穝:");
+			book.get().setPrice(price);
+			bookStoredao.save(book.get());
+			messageList.messagelist.add("ID:"+book.get().getId());
+			messageList.messagelist.add(":"+book.get().getName());
+			messageList.messagelist.add(":"+book.get().getWriter());
+			messageList.messagelist.add("基:"+book.get().getPrice());
+			messageList.messagelist.add("畐:"+book.get().getStorage());
+		}
+		return messageList;
+	
+	}
+	
 }
